Fix OSK (#132)
The OSK appearing affects the layout of main app windows. But the checks for which windows to re-layout was affecting the OSK as well.

Visibly, the OSK was appearing at the top of the output

(cherry picked from commit 27e46daaf21f92ab091c3cf0f27b87f7a8c9c7e1)